Created build sample for windows resource compile
Issue [#173](https://github.com/gradle/gradle-native/issues/173) Windows Resource Support. A sample was needed using
current model instead of deprecated software model.  The build
sample makes use of the 'cpp-application' plugin and wires the
WindowsResourceCompile task for execution.  It is understood
the DSL may change as the plugin may directly support the WRC
task in future releases.

In order for this sample to work, it depends on PR #7432 where
changes were made to the WindowsResourceCompile task to
support the provider api ListProperty<String>.

Signed-off-by: Kent Fletcher <kfletcher@sumglobal.com>